Forge Port
Optimize VarInt writing

While all cases of VarInt writing are optimized, we pay special attention to the 1/2-byte VarInt cases as they are the most common. This has been corroborated by determining the counts of the byte sizes of each VarInt, and it's dramatically skewed towards smaller VarInts. It pays!

fix build